civ() Class Declaration and Global Variables
This is a class declaration for a civilization, civ().
You can specify values by saying civ(Arms_Deployment_Speed = 0.05), etc. Otherwise, the default is our civilization, which can launch interstellar weapons at 0.00025 times the speed of light, ranks 0.7 on the Kardashev scale, and has an 50% aggression score. The aggression score and especially the Communication constant (currently 1) are open to change, as they are arbitrary variables. I'd make an argument that we are a good benchmark for an average level of aggressiveness in a civilization at our level of development, or at least our sample size of 1 makes it difficult to make this value anything other than 0.5 (for now).